requirements.txt
crewai
langchain-openai
python-dotenv
requests
langchain-anthropic 
anthropic